backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.2.2
pandas==2.0.3
python-dotenv==1.0.1
joblib==1.3.2
requests==2.31.0
pytest==8.0.0
matplotlib==3.7.1
seaborn==0.12.2
gunicorn==20.1.0
mlflow==2.8.1
scipy==1.10.1
psycopg2-binary==2.9.9
sqlalchemy==1.4.41
sqlalchemy-utils==0.41.1
alembic==1.12.1